Support XML export
@snoyberg Any feedback welcome. :-)

I'm not sure whether any additional flourishments are necessary for
the XML document. I'm not sure whether optional fields are typically
excluded, left as self-closing tags <group/> or empty <group></group>,
in the XML world.

If there's already an equivalent of `with' lying around, I'd be happy
to use that.

I noticed that while its type is effectively a Producer, I can't use
Producer because it's a rank-N type which becomes concrete when the
attempt to unify it with toXmlRows's generic `row' type variable. I
might be mis-using the conduit abstraction here.